Expand the click areas in the main channel list
A follow-up to [my previous PR](https://github.com/aaronpk/Monocle/pull/25), this time making the entire entry clickable in the main page channels list. This PR changes the click area in the channel list so that the entire area for a channel can be clicked rather than just the name. The changes in the PR result in no visual changes, everything looks exactly as it did before.